web_backend/requirements.txt
fastapi
mangum
upstash-redis
python-multipart
python-dotenv
selenium
webdriver-manager